add more and clear Trim function support
