+++ gse.go
@@ -19,9 +19,6 @@ Package gse Go efficient text segmentation, Go 语言高性能分词
 package gse
 
 import (
-	"unicode"
-	"unicode/utf8"
-
 	"github.com/go-ego/gse/hmm"
 )
 
@@ -163,137 +160,6 @@ func (seg *Segmenter) PosStr(str []SegPos, separator ...string) (r string) {
 			r += add
 		} else {
 			r += add + sep
-		}
-	}
-
-	return
-}
-
-// TrimPunct trim SegPos not space and punct
-func (seg *Segmenter) TrimPunct(se []SegPos) (re []SegPos) {
-	for i := 0; i < len(se); i++ {
-		if !seg.IsStop(se[i].Text) {
-			if se[i].Text != "" {
-				se[i].Text = FilterEmoji(se[i].Text)
-				if len(se[i].Text) > 0 {
-					ru := []rune(se[i].Text)[0]
-					if !unicode.IsSpace(ru) && !unicode.IsPunct(ru) {
-						re = append(re, se[i])
-					}
-				}
-			}
-		}
-	}
-
-	return
-}
-
-// TrimPos trim some pos
-func (seg *Segmenter) TrimPos(se []SegPos, pos ...string) (re []SegPos) {
-	for h := 0; h < len(pos); h++ {
-		if h > 0 {
-			se = re
-			re = nil
-		}
-
-		for i := 0; i < len(se); i++ {
-			if se[i].Pos != pos[h] {
-				re = append(re, se[i])
-			}
-		}
-	}
-
-	return
-}
-
-func notPunct(ru []rune) bool {
-	for i := 0; i < len(ru); i++ {
-		if !unicode.IsSpace(ru[i]) && !unicode.IsPunct(ru[i]) {
-			return true
-		}
-	}
-
-	return false
-}
-
-// Trim trim []string exclude space and punct
-func (seg *Segmenter) Trim(s []string) (r []string) {
-	for i := 0; i < len(s); i++ {
-		if !seg.IsStop(s[i]) {
-			s[i] = FilterEmoji(s[i])
-			ru := []rune(s[i])
-			if len(ru) > 0 {
-				r0 := ru[0]
-				if !unicode.IsSpace(r0) && !unicode.IsPunct(r0) {
-					r = append(r, s[i])
-				} else if len(ru) > 1 && notPunct(ru) {
-					r = append(r, s[i])
-				}
-			}
-		}
-	}
-
-	return
-}
-
-// CutTrim cut string and tirm
-func (seg *Segmenter) CutTrim(str string, hmm ...bool) []string {
-	s := seg.Cut(str, hmm...)
-	return seg.Trim(s)
-}
-
-// PosTrim cut string pos and trim
-func (seg *Segmenter) PosTrim(str string, search bool, pos ...string) []SegPos {
-	p := seg.Pos(str, search)
-	p = seg.TrimPos(p, pos...)
-	return seg.TrimPunct(p)
-}
-
-// PosTrimArr cut string return pos.Text []string
-func (seg *Segmenter) PosTrimArr(str string, search bool, pos ...string) (re []string) {
-	p1 := seg.PosTrim(str, search, pos...)
-	for i := 0; i < len(p1); i++ {
-		re = append(re, p1[i].Text)
-	}
-
-	return
-}
-
-// PosTrimStr cut string return pos.Text string
-func (seg *Segmenter) PosTrimStr(str string, search bool, pos ...string) string {
-	pa := seg.PosTrimArr(str, search, pos...)
-	return seg.CutStr(pa)
-}
-
-// FilterEmoji filter the emoji
-func FilterEmoji(content string) (new string) {
-	for _, value := range content {
-		_, size := utf8.DecodeRuneInString(string(value))
-		if size <= 3 {
-			new += string(value)
-		}
-	}
-
-	return
-}
-
-// FilterSymbol filter the symbol
-func FilterSymbol(text string) (new string) {
-	for _, value := range text {
-		if !unicode.IsSymbol(value) &&
-			!unicode.IsSpace(value) && !unicode.IsPunct(value) {
-			new += string(value)
-		}
-	}
-
-	return
-}
-
-// FilterLang filter the language
-func FilterLang(text, lang string) (new string) {
-	for _, value := range text {
-		if unicode.IsLetter(value) || unicode.Is(unicode.Scripts[lang], value) {
-			new += string(value)
 		}
 	}
